requirements.txt
flask
flask-socketio
flask-cors
python-dotenv
eventlet
markupsafe
